Refactor: WAT: Merge classes WATVisitor and WATGenerator
